chore: Headerコンポーネントを共通コンポーネントとして切り出した
Why: 単体のテストを行いため
