+++ apps/web/src/app/AppLayout/AppLayout.module.css
@@ -10,7 +10,3 @@
   flex-grow: 1;
   padding: var(--space-2);
 }
-
-.header {
-  padding: var(--space-2);
-}
