+++ apps/web/src/app/AppLayout/AppLayout.tsx
@@ -1,8 +1,8 @@
-import { HamburgerMenuIcon } from "@radix-ui/react-icons"
-import { Flex, IconButton } from "@radix-ui/themes"
+import { Flex } from "@radix-ui/themes"
 import { Outlet } from "react-router-dom"
 import { SidebarTreeButton } from "../../components/buttons/SidebarTreeButton"
 import { useAuthCheck } from "../../utils/auth/useAuthCheck"
+import { Header } from "../Header"
 import { Sidebar, useSidebar } from "../Sidebar"
 import styles from "./AppLayout.module.css"
 import { getPaymentsSidebarTree } from "./getPaymentsSidebarTree"
@@ -29,19 +29,5 @@ export function AppLayout() {
         </main>
       </Flex>
     </Flex>
-  )
-}
-
-interface HeaderProps {
-  onMenuClick: () => void
-}
-
-function Header({ onMenuClick }: HeaderProps) {
-  return (
-    <header className={styles.header}>
-      <IconButton variant="surface" onClick={onMenuClick}>
-        <HamburgerMenuIcon width="18" height="18" />
-      </IconButton>
-    </header>
   )
 }
