+++ apps/web/src/app/Header/Header.module.css
@@ -1,0 +1,3 @@
+.header {
+  padding: var(--space-2);
+}
